+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## bliss 0.12.0
 * Fix some old TODOs.
 * Add a full pipeline benchmark.
+* Add some optimizations for speed.
 
 ## bliss 0.11.4
 * Bump dependencies, most notably updating the symphonia module after